requirements.txt
Flask==2.0.1
Flask_RESTful==0.3.9
requests==2.26.0
Pillow==8.3.2
